nokia migrate

+++ src/Services/Nokia/NokiaClient.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Services.MyHealth.Domain;
+using Repositories.Models;
 using Services.Nokia.Domain;
 using Utils;
 
@@ -42,7 +42,7 @@ namespace Services.Nokia
                     weights.Add(new Weight
                     {
                         DateTime = weightMeasure.date.ToDateFromUnixTime(),
-                        Kg = weightMeasure.measures.First(x => x.type == WeightKgMeasureTypeId).value * Math.Pow(10, weightMeasure.measures.First(x => x.type == WeightKgMeasureTypeId).unit),
+                        Kg = (Decimal)( weightMeasure.measures.First(x => x.type == WeightKgMeasureTypeId).value * Math.Pow(10, weightMeasure.measures.First(x => x.type == WeightKgMeasureTypeId).unit)),
 
                         //todo set if available
                         //  FatRatioPercentage = bodyMeasuregrp.measures.First(x => x.type == FatRatioPercentageMeasureTypeId).value * Math.Pow(10, bodyMeasuregrp.measures.First(x => x.type == FatRatioPercentageMeasureTypeId).unit),
